Fixed sonar

+++ src/main/java/g3701_3800/s3709_design_exam_scores_tracker/ExamTracker.java
@@ -19,7 +19,7 @@ public class ExamTracker {
     public void record(int time, int score) {
         ti.add(time);
         long pv = pr.isEmpty() ? 0L : pr.get(pr.size() - 1);
-        pr.add(pv + (long) score);
+        pr.add(pv + score);
     }
 
     public long totalScore(int startTime, int endTime) {
